semester-3/python-iii/flask-231206/pyproject.toml
[project]
name = "mysqlexample"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "pan93412", email = "pan93412@gmail.com" }
]
dependencies = [
    "sqlalchemy>=2.0.23",
    "mysql-connector-python>=8.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.11.0",
]

[tool.hatch.metadata]
allow-direct-references = true
